frontend + backend : Simulation of envoi + checking if user is logged in or NOT + calculating the total price of the envoi and volume of the envoi = OK
Took 3 hours 20 minutes

+++ backend/src/main/java/com/moncolisapp/backend/dto/ColisDTO.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * DTO for {@link com.moncolisapp.backend.entities.Colis}
@@ -15,9 +16,10 @@ import java.io.Serializable;
 @NoArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class ColisDTO implements Serializable {
-    private double hauteur;
-    private double largeur;
-    private double longueur;
-    private double poidsColis;
-    private double volumeColis;
+    private BigDecimal hauteur;
+    private BigDecimal largeur;
+    private BigDecimal longueur;
+    private BigDecimal poidsColis;
+    private BigDecimal volumeColis;
+
 }
